Bug 1047168 - Change test_browser_play_youtube_video.py to run local content and embedded video
