+++ tests/python/gaia-ui-tests/gaiatest/tests/functional/browser/manifest.ini
@@ -22,7 +22,10 @@ smoketest = true
 
 [test_browser_bookmark.py]
 
-[test_browser_play_youtube_video.py]
+[test_browser_play_embedded_3gp_video.py]
 smoketest = true
-# Bug 877594 - Unable to play YouTube video in B2G desktop client
-fail-if = device == "desktop"
+# Bug 1050271 - 3GP file in a video tag will not play on Linux using desktopb2g
+#skip-if = os == "linux" && device == "desktop"
+
+[test_browser_play_embedded_mp4_video.py]
+smoketest = true
