+++ tests/python/gaia-ui-tests/gaiatest/tests/tbpl-manifest.ini
@@ -39,6 +39,12 @@ disabled = Bug 1022132 Intermittent Failure on TBPL
 [functional/browser/test_browser_bookmark.py]
 [functional/browser/test_browser_lan.py]
 [functional/browser/test_browser_navigation.py]
+[functional/browser/test_browser_play_embedded_3gp_video.py]
+# Bug 1050271 - 3GP file in a video tag will not play on Linux using desktopb2g
+# skip-if = os == "linux"
+[functional/browser/test_browser_play_embedded_mp4_video.py]
+# Bug 1049463 TBPL cannot play the mp4 codec on Mac
+# skip-if = os == "mac"
 [functional/browser/test_browser_tabs.py]
 [functional/calendar/test_calendar_flick_through_months.py]
 [functional/calendar/test_calendar_new_event_appears_on_all_calendar_views.py]
